feat(ssr): prefetch server-side de milestones via cookie (PM-171 Fase 1)
Re-aplicación de Fase 1 sobre el dev actual (el equipo avanzó ~20 commits y le
añadió a milestones un feature de progreso real por tareas + link "Ver proyecto";
todo eso se preserva intacto). Apila sobre Fase 0 (ya en dev). SEGURO y NO
desplegable hasta que Fase 0 esté en prod: no incluye middleware enforcing.

Progressive enhancement: SSR si hay cookie pm_session, fallback al render cliente
(localStorage) si no, así no rompe ninguna sesión.

- src/lib/serverAuth.ts (nuevo): getServerToken() lee pm_session via cookies();
  serverFetch() GET autenticado server-side, per-request (sin cache de modulo) y
  con timeout 2.5s + AbortController para no bloquear el TTFB ante cold-starts.
- milestones migrado a Server Component (force-dynamic): page.tsx prefetcha el
  overview y pasa initialProjects a MilestonesClient (use client). _mappers.ts
  extrae los mappers puros. El feature del equipo (getProjectTasks por proyecto
  seleccionado, sprintProgressById, "Ver proyecto") queda intacto client-side.
- report route: cookie pm_session como FALLBACK del token de query (primario).
- userService: contador de generacion (no repuebla cache tras invalidate) +
  limpieza condicional del inflight; invalidateUsersCache() en logout.
