+++ src/components/auth/AuthProvider.tsx
@@ -18,7 +18,7 @@ import {
   syncSessionCookie,
 } from '@/lib/auth'
 import { AuthContext, AuthContextValue } from '@/hooks/useAuth'
-import { getUserProfileDetails } from '@/services/userService'
+import { getUserProfileDetails, invalidateUsersCache } from '@/services/userService'
 
 /*
 skill/área/foto NO vienen en /auth/me. La única fuente del backend es GET /users
@@ -131,6 +131,9 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
 
   function logout() {
     logoutUser()
+    // No servir el directorio cacheado del usuario anterior tras logout+login
+    // sin recargar (la caché de getUsersRaw no incluye identidad en su key).
+    invalidateUsersCache()
     setUser(null)
     setIsAuthenticated(false)
   }
